requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2023.7.22
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==2.3.3
frozenlist==1.5.0
gigachat==0.1.39
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-texttospeech==2.25.0
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-core==0.3.35
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langsmith==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.2
multidict==6.1.0
numpy==1.26.4
openai==1.3.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
-e git+ssh://git@github.com/wowlocal/ssml-voice.git@ac4b77be2d965f4a248e5707131887b93edb17a1#egg=voice
websockets==15.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
